Added enzyme tests to a few components in the company folder WIP
Wrote basic unit test for CompanyCard component. Snapshot test with hard-coded data passed as a prop

Wrote test for CompanyDisplay component. Checks that loading text is displayed in the component with a shallow mount,
Second test has jest spying on componentDidMount lifecycle method, then mocking an API call and checking that the component renders the dummy-data